+++ packages/staker/src/cli.js
@@ -22,17 +22,17 @@ import {
 import {
   skillEth2Stats,
   skillBeaconLastBlock,
-  skillGetValidatorBlance,
+  skillGetValidatorBalance,
   skillsValidator,
 } from './eth2/saihubot-cli-skill-eth2';
 // import {skills as skillsDefi} from './defi/saihubot-cli-skill-defi';
 import {skillsGas} from './eth/saihubot-cli-skill-ethgas';
-import {skillGetBlance, skillGetEthBlance} from './saihubot-cli-skill-balance';
+import {skillGetBalance, skillGetEthBalance} from './saihubot-cli-skill-balance';
 import {skills as skillsXDAI} from './xdai/saihubot-cli-skill-xdai';
 import {skills as skillsBSC} from './bsc/saihubot-cli-skill-bsc';
 import {skills as skillsMatic} from './matic/saihubot-cli-skill-matic';
 import {skills as skillsZkSync} from './zksync/saihubot-cli-skill-zksync';
-import {skills as skillsOptimism} from './zksync/saihubot-cli-skill-optimism';
+import {skills as skillsOptimism} from './optimism/saihubot-cli-skill-optimism';
 import {skills as skillsChain} from './saihubot-cli-skill-chainid';
 import {skillAwesomeEth} from './saihubot-skill-awesome';
 
@@ -67,12 +67,12 @@ const bot = new SaihuBot({
     addonFetchGas,
   ],
   skills: [
-    skillGetBlance,
-    skillGetEthBlance,
+    skillGetBalance,
+    skillGetEthBalance,
     // ...skillsDefi,
     skillLastBlock,
     ...skillsAddress,
-    skillGetValidatorBlance,
+    skillGetValidatorBalance,
     skillBeaconLastBlock,
     ...skillsValidator,
     skillEth2Stats,