+++ packages/staker/src/matic/saihubot-cli-skill-matic.js
@@ -1,7 +1,11 @@
 'use strict';
+import React from 'react';
+import { t } from 'saihubot-cli-adapter/dist/i18n';
 
-import {getConfig, singleAddr, addrTxSearch} from '../utils';
+import {getConfig, parseArg, singleAddr, toArray, addrTxSearch} from '../utils';
 import {i18nAddr} from '../i18n';
+import MaticBalances from './MaticBalances';
+import {maticFetch} from './utils'
 
 /**
  * Check address or tx on Matic.
@@ -41,7 +45,38 @@ export const skillSearchMatic = {
   },
 };
 
+
+/**
+ * Get balance of [addr] on MATIC Chain.
+ *
+ * can pass the address, or pre-define the
+ * SAIHUBOT_ADDR environment variable
+ */
+ export const skillGetMaticBalance = {
+  name: 'balance-matic',
+  help: '💰balance-matic - Show address balance on MATIC chain',
+  requirements: {
+    addons: ['fetch'],
+  },
+  rule: /(^balance-matic )(.*)|^balance-matic$/i,
+  action: function(robot, msg) {
+    let addr = '';
+    if (msg[2] === undefined) {
+      addr = getConfig('ADDR', '');
+      if (!addr) {
+        robot.send(t('needAddr', {i18n: i18nBalance}));
+        robot.render();
+        return;
+      }
+    }
+    const addrs = toArray(addr || parseArg(msg[2]));
+    robot.sendComponent(<MaticBalances addresses={addrs} fetch={robot.addons.fetch} maticFetch={maticFetch} />);
+    robot.render();
+  },
+}
+
 const skills = [
+  skillGetMaticBalance,
   skillSearchMatic,
 ];
 export {skills};