+++ packages/staker/src/matic/token.js
@@ -1,0 +1,87 @@
+// https://docs.matic.network/docs/develop/network-details/mapped-tokens
+export const TOKENMAP = [
+  {
+    address: '0x7ceB23fD6bC0adD59E62ac25578270cFf1b9f619',
+    name: 'WETH',
+    symbol: 'WETH',
+    decimals: 18,
+  },
+  {
+    address: '0x2791Bca1f2de4661ED88A30C99A7a9449Aa84174',
+    name: 'USDC',
+    symbol: 'USDC',
+    decimals: 6,
+  },
+  {
+    address: '0x8505b9d2254A7Ae468c0E9dd10Ccea3A837aef5c',
+    name: 'Compound',
+    symbol: 'COMP',
+    decimals: 18,
+  },
+  {
+    address: '0x313d009888329C9d1cf4f75CA3f32566335bd604',
+    name: 'Lend',
+    symbol: 'LEND',
+    decimals: 18,
+  },
+  {
+    address: '0xDA537104D6A5edd53c6fBba9A898708E465260b6',
+    name: 'YFI',
+    symbol: 'YFI',
+    decimals: 18,
+  },
+  {
+    address: '0xc2132D05D31c914a87C6611C10748AEb04B58e8F',
+    name: 'USDT',
+    symbol: 'USDT',
+    decimals: 6,
+  },
+  {
+    address: '0x8f3Cf7ad23Cd3CaDbD9735AFf958023239c6A063',
+    name: 'DAI',
+    symbol: 'DAI',
+    decimals: 18,
+  },
+  {
+    address: '0xdAb529f40E671A1D4bF91361c21bf9f0C9712ab7',
+    name: 'BUSD',
+    symbol: 'BUSD',
+    decimals: 18,
+  },
+  {
+    address: '0xA1c57f48F0Deb89f569dFbE6E2B7f46D33606fD4',
+    name: 'MANA',
+    symbol: 'MANA',
+    decimals: 18,
+  },
+  {
+    address: '0xA1c57f48F0Deb89f569dFbE6E2B7f46D33606fD4',
+    name: 'WBTC',
+    symbol: 'WBTC',
+    decimals: 8,
+  },
+  {
+    address: '0x71B821aa52a49F32EEd535fCA6Eb5aa130085978',
+    name: '0xBTC',
+    symbol: '0xBTC',
+    decimals: 8,
+  },
+  {
+    address: '0x578360AdF0BbB2F10ec9cEC7EF89Ef495511ED5f',
+    name: 'KIWI',
+    symbol: 'KIWI',
+    decimals: 8,
+  },
+  {
+    address: '0x556f501CF8a43216Df5bc9cC57Eb04D4FFAA9e6D',
+    name: 'DUST',
+    symbol: 'DUST',
+    decimals: 8,
+  },
+  {
+    address: '0x556f501CF8a43216Df5bc9cC57Eb04D4FFAA9e6D',
+    name: 'DUST',
+    symbol: 'DUST',
+    decimals: 8,
+  },
+];