+++ packages/staker/src/optimism/saihubot-cli-skill-optimism.js
@@ -77,7 +77,7 @@ export const skillSearchOptimism = {
 // }
 
 const skills = [
-  // skillGetOptimismBlance,
+  // skillGetOptimismBalance,
   skillSearchOptimism,
 ];
 export {skills};